gov_track_hk_web/requirements.txt
django==1.11.11
setuptools
MarkupSafe==0.23
mysqlclient==1.3.12
Pillow==3.3.1
PyPDF2==1.26.0
SQLAlchemy==1.0.14
alembic==0.8.6
argparse==1.2.1
cssselect==0.9.2
cycler==0.10.0
demjson==2.2.4
django-dirtyfields==1.0
djangorestframework==3.7.7
jieba==0.38
lxml==3.6.0
numpy==1.11.1
pandas==0.18.1
pyparsing==2.1.8
pyquery==1.2.13
pytagcloud==0.3.5
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.6
requests==2.10.0
scraperwiki==0.5.1
simplejson==3.8.2
six==1.10.0
wordcloud==1.2.1
PyExecJS==1.4.0
django-webpack-loader==0.6.0
django-letsencrypt==3.0.0
pylama==7.4.3
isort==4.3.4
